udpated

+++ binary_trees.h
@@ -43,5 +43,5 @@ void binary_tree_inorder(const binary_tree_t *tree, void (*func)(int));
 void binary_tree_postorder(const binary_tree_t *tree, void (*func)(int));
 size_t binary_tree_height(const binary_tree_t *tree);
 
-size_t max(size_t a, size_t b);
+size_t max(int a, int b);
 #endif /*_BINARY_TREES_H_*/
